ryu/ryu/app/files/requirements.txt
mininet>=2.3.0
psycopg2-binary>=2.9.0
ryu>=4.34
eventlet>=0.33.0
